# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Carrier Performance Scorecard API
  version: 0.1.0
  description: Public demo API for fictional carrier performance scorecards. The API is read only except for the demo QBR brief generator.
servers:
  - url: http://127.0.0.1:3100
    description: Local development
paths:
  /api/health:
    get:
      summary: Read service health and safe dependency status.
      responses:
        "200":
          description: Health response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/demo-data:
    get:
      summary: Read safe demo dataset metadata and row counts.
      responses:
        "200":
          description: Demo data metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoDataResponse"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/scorecards/options:
    get:
      summary: Read filter options for carriers and periods.
      responses:
        "200":
          description: Filter options.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilterOptionsResponse"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/scorecards/summary:
    get:
      summary: Read carrier scorecard summaries for optional filters.
      parameters:
        - $ref: "#/components/parameters/CarrierId"
        - $ref: "#/components/parameters/Region"
        - $ref: "#/components/parameters/ProductType"
        - $ref: "#/components/parameters/Period"
      responses:
        "200":
          description: Scorecard summary response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScorecardsSummaryResponse"
        "400":
          $ref: "#/components/responses/InvalidFilter"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/carriers/{carrierId}/scorecard:
    get:
      summary: Read detailed scorecard data for one carrier.
      parameters:
        - name: carrierId
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/Region"
        - $ref: "#/components/parameters/ProductType"
        - $ref: "#/components/parameters/Period"
      responses:
        "200":
          description: Carrier scorecard response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarrierScorecardResponse"
        "400":
          $ref: "#/components/responses/InvalidFilter"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/evidence:
    get:
      summary: Read evidence items for filters, dimensions, delay reasons, or selected evidence IDs.
      parameters:
        - $ref: "#/components/parameters/CarrierId"
        - $ref: "#/components/parameters/Region"
        - $ref: "#/components/parameters/ProductType"
        - $ref: "#/components/parameters/Period"
        - name: dimension
          in: query
          schema:
            type: string
        - name: delayReason
          in: query
          schema:
            type: string
        - name: evidenceIds
          in: query
          description: Comma separated RFC4122 UUID values.
          schema:
            type: string
        - name: cap
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Evidence response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceResponse"
        "400":
          $ref: "#/components/responses/InvalidFilter"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/qbr/brief:
    post:
      summary: Generate a safe demo QBR brief for one carrier.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QbrBriefRequest"
      responses:
        "200":
          description: Generated QBR brief.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QbrBriefResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/ServerError"
components:
  parameters:
    CarrierId:
      name: carrierId
      in: query
      schema:
        type: string
      description: Optional carrier ID filter.
    Region:
      name: region
      in: query
      schema:
        type: string
      description: Optional region filter.
    ProductType:
      name: productType
      in: query
      schema:
        type: string
      description: Optional product type filter.
    Period:
      name: period
      in: query
      schema:
        type: string
      description: Optional period seed key filter.
  responses:
    InvalidFilter:
      description: Invalid filter input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidRequest:
      description: Invalid request body or unsupported input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Controlled server error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          const: false
        error:
          type: object
          required: [message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    HealthResponse:
      type: object
      required: [ok, service, demoMode, time, dependencies]
      properties:
        ok:
          const: true
        service:
          type: string
        demoMode:
          type: boolean
        time:
          type: string
          format: date-time
        dependencies:
          type: object
          required: [databaseConfigured]
          properties:
            databaseConfigured:
              type: boolean
    DemoDataResponse:
      type: object
      required: [ok, demoMode, dataset, counts, message]
      properties:
        ok:
          type: boolean
        demoMode:
          type: boolean
        dataset:
          anyOf:
            - $ref: "#/components/schemas/DatasetMetadata"
            - type: "null"
        counts:
          type: object
          properties:
            carriers:
              type: integer
            periods:
              type: integer
            deliveryRecords:
              type: integer
            evidenceItems:
              type: integer
        message:
          type:
            - string
            - "null"
    DatasetMetadata:
      type: object
      properties:
        id:
          type: string
        seedVersion:
          type: string
        fingerprint:
          type: string
        seededAt:
          type: string
          format: date-time
    FilterOptionsResponse:
      type: object
      required: [ok, carriers, periods]
      properties:
        ok:
          const: true
        carriers:
          type: array
          items:
            $ref: "#/components/schemas/CarrierOption"
        periods:
          type: array
          items:
            $ref: "#/components/schemas/PeriodOption"
    CarrierOption:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        shortCode:
          type: string
        relationshipTier:
          type: string
        regionFocus:
          type: string
    PeriodOption:
      type: object
      properties:
        seedKey:
          type: string
        label:
          type: string
        startDate:
          type: string
        endDate:
          type: string
    ScorecardsSummaryResponse:
      type: object
      description: High level scorecard list and aggregate context.
      additionalProperties: true
    CarrierScorecardResponse:
      type: object
      description: Detailed scorecard, trend, dimension, and evidence context for one carrier.
      additionalProperties: true
    EvidenceResponse:
      type: object
      description: Evidence list response with safe fictional evidence only.
      additionalProperties: true
    QbrBriefRequest:
      type: object
      required: [carrierId]
      additionalProperties: false
      properties:
        carrierId:
          type: string
          maxLength: 80
        region:
          type:
            - string
            - "null"
          maxLength: 120
        productType:
          type:
            - string
            - "null"
          maxLength: 120
        period:
          type:
            - string
            - "null"
          maxLength: 20
        variant:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 9
    QbrBriefResponse:
      type: object
      description: Demo safe generated brief response.
      additionalProperties: true